requirements.txt
absl-py==0.9.0
alabaster==0.7.12
argh==0.26.2
asn1crypto==1.3.0
astor==0.7.1
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==3.1.4
blinker==1.4
blis==0.4.1
cachetools==3.1.1
catalogue==1.0.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
click-plugins==1.1.1
cligj==0.5.0
cloudpickle==1.3.0
colorama==0.4.3
comtypes==1.1.7
convertdate==2.1.3
cryptography==2.8
cycler==0.10.0
cymem==2.0.3
Cython==0.29.16
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
docutils==0.16
en-core-web-sm==2.2.5
entrypoints==0.3
ephem==3.7.7.1
fbprophet==0.6
Fiona==1.8.13
flake8==3.7.9
Flask==1.1.1
Flask-Cors==3.0.8
future==0.18.2
gast==0.2.2
GDAL==3.0.4
geographiclib==1.50
geopandas==0.7.0
geopy==1.22.0
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
gTTS==2.1.1
gTTS-token==1.1.3
h5py==2.10.0
holidays==0.10.1
idna==2.9
imagesize==1.2.0
importlib-metadata==1.5.0
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==21.1.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
LunarCalendar==0.0.9
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
mkl-service==2.3.0
munch==2.5.0
murmurhash==1.0.0
nbconvert==5.6.1
nbformat==5.0.4
newsapi-python==0.2.6
nltk==3.4.4
notebook==6.0.3
numpy==1.18.1
numpydoc==0.9.2
oauthlib==3.0.1
olefile==0.46
opt-einsum==0+untagged.56.g2664021.dirty
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.5.2
pathlib-mate==1.0.0
pathtools==0.1.2
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
plac==0.9.6
playsound==1.2.2
plotly==4.6.0
pluggy==0.13.1
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.11.4
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyAudio==0.2.11
pycodestyle==2.5.0
pycparser==2.20
pydocstyle==4.0.1
pydub==0.23.1
pyflakes==2.1.1
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyproj==2.6.1.post1
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtWebEngine==5.12.1
pyreadline==2.1
pyrsistent==0.16.0
PySocks==1.7.1
pystan==2.19.0.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.9
pyttsx3==2.87
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==18.1.1
QDarkStyle==2.8
QtAwesome==0.7.0
qtconsole==4.7.2
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.2.0
retrying==1.3.3
rope==0.16.0
rsa==4.0
Rtree==0.9.3
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
Shapely==1.7.0
six==1.14.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
soupsieve==2.0
spacy==2.2.3
SpeechRecognition==3.6.3
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==4.1.3
spyder-kernels==1.9.1
SQLAlchemy==1.3.17
srsly==1.0.0
statsmodels==0.11.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
textblob==0.15.3
thinc==7.3.0
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
tweepy==3.8.0
ujson==1.35
urllib3==1.25.8
uszipcode==0.2.4
wasabi==0.6.0
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.16.1
win-inet-pton==1.1.0
wincertstore==0.2
wordcloud==1.6.0
wrapt==1.12.1
yapf==0.28.0
zipp==2.2.0
